# Faster AbstractArray hashing