@mbauman 2021-03-08
No, I actually didn't realize that `hash(AbstractArray)` isn't stable across sessions.  IIRC I just thought it was a cute way of avoiding the need to come up with that random number (and 32-bit truncation).