@mbauman 2021-03-08
> Actually I might be wrong about that, I assumed it came from the pointer.

Not sure what's happening, but it seems to be preserved within a given precompile/sysimg.