@oscardssmith 2021-03-08
While we're speeding this up, would it make sense to change the behavior of hashing small arrays (<25 elements say) to hash all elements rather than `log(n)`? I'd imagine that this would likely be faster (at least for `Array`)