@timholy 2021-03-09
If that's viable I'd be fine with it, but it's a different kind of change than the one here and should be a separate PR.